Move pipe_cell_list to separate file
The pipe_cell_list is not really part of the canvas, so refactor it into
a separate file.
